Restructure file tree and resolve jest failures
