+++ src/reducer/replace/replace.js
@@ -1,6 +1,6 @@
-import plugins from '../plugins/plugins';
-import TreeParser from '../tree-parser/tree-parser';
-import ExampleGenerator from '../generators/example/example-generator';
+import plugins from '../../plugins/plugins';
+import TreeParser from '../../tree-parser/tree-parser';
+import ExampleGenerator from '../../generators/example/example-generator';
 
 /**
  * Replaces a component with another component.
@@ -26,5 +26,3 @@ const replace = (state, action) => {
 };
 
 export default replace;
-
-// replace, remove, insert
